+++ packages/adapter-node/index.js
@@ -1,11 +1,18 @@
-import { readFileSync, statSync, createReadStream, createWriteStream, writeFileSync } from 'fs';
+import esbuild from 'esbuild';
+import {
+	createReadStream,
+	createWriteStream,
+	existsSync,
+	readFileSync,
+	statSync,
+	writeFileSync
+} from 'fs';
 import { join } from 'path';
-import { fileURLToPath } from 'url';
 import { pipeline } from 'stream';
+import glob from 'tiny-glob';
+import { fileURLToPath } from 'url';
 import { promisify } from 'util';
 import zlib from 'zlib';
-import esbuild from 'esbuild';
-import glob from 'tiny-glob';
 
 const pipe = promisify(pipeline);
 
@@ -56,6 +63,7 @@ export default function ({
 				format: 'esm',
 				platform: 'node',
 				target: 'node12',
+				inject: [join(files, 'shims.js')],
 				define: {
 					esbuild_app_dir: '"' + config.kit.appDir + '"'
 				}
@@ -65,7 +73,7 @@ export default function ({
 			await utils.prerender({
 				dest: `${out}/prerendered`
 			});
-			if (precompress) {
+			if (precompress && existsSync(`${out}/prerendered`)) {
 				utils.log.minor('Compressing prerendered pages');
 				await compress(`${out}/prerendered`);
 			}