+++ packages/kit/src/core/config/index.spec.js
@@ -36,7 +36,11 @@ test('fills in defaults', () => {
 				files: {
 					include: ['**'],
 					exclude: []
-				}
+				},
+				emitTypes: true
+			},
+			serviceWorker: {
+				exclude: []
 			},
 			paths: {
 				base: '',
@@ -131,7 +135,11 @@ test('fills in partial blanks', () => {
 				files: {
 					include: ['**'],
 					exclude: []
-				}
+				},
+				emitTypes: true
+			},
+			serviceWorker: {
+				exclude: []
 			},
 			paths: {
 				base: '',